DAS-2232 - added a module to contain coordinate methods
